fix image manifest filename
since mickledore the suffix is not always the MACHINE variable. It can be set with another variable IMAGE_MACHINE_SUFFIX, which defaults to "-${MACHINE}".

this can be used to cleanup the deploy folder (since the machine information is already part of the folder structure leading to the images).

without this change building using the st-partitions-image.bbclass will fail